modularized everything, adjusted and removed unnecessary conditionals/variables from handle_shoot and increased ball size to 10
