+++ screen_settings.py
@@ -1,0 +1,10 @@
+import pygame
+pygame.init()
+
+WIDTH = 1200
+HEIGHT = 700
+MID_W = WIDTH // 2
+MID_H = HEIGHT // 2
+REL_SIZE = 40
+screen = pygame.display.set_mode((WIDTH, HEIGHT))
+pygame.display.set_caption("Triangular Assault")
