+++ sfx.py
@@ -1,0 +1,7 @@
+import pygame
+pygame.init()
+
+hit_sound_effect = pygame.mixer.Sound('assets/bounce.wav')
+wave_sound_effect = pygame.mixer.Sound('assets/scoring-sound.wav')
+damage_sound_effect = pygame.mixer.Sound('assets/explode.ogg')
+shoot_sound_effect = pygame.mixer.Sound('assets/s2297_from_gd.ogg')
