+++ styles.py
@@ -1,0 +1,12 @@
+import pygame
+
+pygame.init()
+
+# colors
+WHITE = (255, 255, 255)
+BLACK = (0, 0, 0)
+
+# fonts
+hud_font = pygame.font.Font('./assets/VCR_OSD_MONO_1.001.ttf', 40)
+screen_font = pygame.font.SysFont('Arial', 30)
+screen_font_bold = pygame.font.SysFont('Arial Bold', 120)
